requirements.txt
python-dotenv==1.0.0
Django==5.1.6
Pillow==10.2.0
django-crispy-forms==2.1.0
crispy-bootstrap5==0.7
django-allauth==0.60.1
django-storages==1.14.2
django-widget-tweaks==1.5.0
django-filter==23.5
django-environ==0.11.2
django-ckeditor==6.7.0
django-humanize==0.1.2
django-import-export==3.3.3
python-dateutil==2.8.2
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
boto3==1.34.52
django-ses==3.5.2
django-recaptcha==3.0.0
django-cors-headers==4.3.1
celery==5.3.6
redis==5.0.1
bleach==6.1.0
Markdown==3.5.2
django-cleanup==8.0.0
